# examples/complex-test-app/docs/api/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "complex-test-app API",
    "version": "1.0.0",
    "description": "Generated by Reqcord from the application's integration tests."
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
...
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "operationId": "post_api_v1_auth_login",
        "summary": "Login",
        "description": "api/v1/auth#login\nRoute: api_v1_auth_login",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "ada@example.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "{{password}}"
                  }
                },
                "required": [
                  "email",
                  "password"
                ]
              },
              "example": {
                "email": "ada@example.com",
                "password": "{{password}}"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string",
                      "example": "{{token}}"
                    },
                    "token_type": {
                      "type": "string",
                      "example": "Bearer"
                    },
                    "expires_in": {
                      "type": "integer",
                      "example": 3600
                    }
                  },
                  "required": [
                    "token",
                    "token_type",
                    "expires_in"
                  ]
                },
                "example": {
                  "token": "{{token}}",
                  "token_type": "Bearer",
                  "expires_in": 3600
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Invalid email or password"
                    }
                  },
                  "required": [
                    "error"
                  ]
                },
                "example": {
                  "error": "Invalid email or password"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/orders": {
      "get": {
        "operationId": "get_api_v1_orders",
        "summary": "List Orders",
        "description": "api/v1/orders#index\nRoute: api_v1_orders",
        "tags": [
          "Orders"
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "shipped"
              ],
              "example": "pending"
            },
            "example": "pending"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer",
                        "example": 2
                      },
                      "user_id": {
                        "type": "integer",
                        "example": 1
                      },
                      "status": {
                        "type": "string",
                        "enum": [
                          "pending",
                          "shipped"
                        ],
                        "example": "shipped"
                      },
                      "currency": {
                        "type": "string",
                        "example": "USD"
                      },
                      "total_cents": {
                        "type": "integer",
                        "example": 3200
                      }
                    },
                    "required": [
                      "id",
                      "user_id",
                      "status",
                      "currency",
                      "total_cents"
                    ]
                  }
                },
                "example": [
                  {
                    "id": 1,
                    "user_id": 1,
                    "status": "pending",
                    "currency": "USD",
                    "total_cents": 2700
                  },
                  {
                    "id": 2,
                    "user_id": 1,
                    "status": "shipped",
                    "currency": "USD",
                    "total_cents": 3200
                  }
                ]
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Unauthorized"
                    }
                  },
                  "required": [
                    "error"
                  ]
                },
                "example": {
                  "error": "Unauthorized"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "operationId": "post_api_v1_orders",
        "summary": "Create Order",
        "description": "api/v1/orders#create",
        "tags": [
          "Orders"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "order": {
                    "type": "object",
                    "properties": {
                      "line_items": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "sku": {
                              "type": "string",
                              "example": "MUG-001"
                            },
                            "quantity": {
                              "type": "integer",
                              "example": 1
                            }
                          },
                          "required": [
                            "sku",
                            "quantity"
                          ]
                        }
                      },
                      "shipping_address": {
                        "type": "object",
                        "properties": {
                          "line1": {
                            "type": "string",
                            "example": "1 Analytical Engine Way"
                          },
                          "city": {
                            "type": "string",
                            "example": "London"
                          },
                          "country": {
                            "type": "string",
                            "enum": [
                              "GB",
                              "IT"
                            ],
                            "example": "GB"
                          }
                        },
                        "required": [
                          "line1",
                          "city",
                          "country"
                        ]
                      }
                    },
                    "required": [
                      "line_items",
                      "shipping_address"
                    ]
                  }
                },
                "required": [
                  "order"
                ]
              },
              "example": {
                "order": {
                  "line_items": [
                    {
                      "sku": "TEA-001",
                      "quantity": 2
                    },
                    {
                      "sku": "MUG-001",
                      "quantity": 1
                    }
                  ],
                  "shipping_address": {
                    "line1": "1 Analytical Engine Way",
                    "city": "London",
                    "country": "GB"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "example": 4
                    },
                    "status": {
                      "type": "string",
                      "example": "pending"
                    },
                    "currency": {
                      "type": "string",
                      "example": "USD"
                    },
                    "total_cents": {
                      "type": "integer",
                      "example": 4800
                    },
                    "line_items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "sku": {
                            "type": "string",
                            "example": "MUG-001"
                          },
                          "name": {
                            "type": "string",
                            "example": "Stoneware Mug"
                          },
                          "quantity": {
                            "type": "integer",
                            "example": 1
                          },
                          "unit_price_cents": {
                            "type": "integer",
                            "example": 2400
                          },
                          "subtotal_cents": {
                            "type": "integer",
                            "example": 2400
                          }
                        },
                        "required": [
                          "sku",
                          "name",
                          "quantity",
                          "unit_price_cents",
                          "subtotal_cents"
                        ]
                      }
                    },
                    "shipping_address": {
                      "type": "object",
                      "properties": {
                        "line1": {
                          "type": "string",
                          "example": "1 Analytical Engine Way"
                        },
                        "city": {
                          "type": "string",
                          "example": "London"
                        },
                        "country": {
                          "type": "string",
                          "enum": [
                            "GB",
                            "IT"
                          ],
                          "example": "GB"
                        }
                      },
                      "required": [
                        "line1",
                        "city",
                        "country"
                      ]
                    }
                  },
                  "required": [
                    "id",
                    "status",
                    "currency",
                    "total_cents",
                    "line_items",
                    "shipping_address"
                  ]
                },
                "example": {
                  "id": 4,
                  "status": "pending",
                  "currency": "USD",
                  "total_cents": 4800,
                  "line_items": [
                    {
                      "sku": "TEA-001",
                      "name": "Earl Grey",
                      "quantity": 2,
                      "unit_price_cents": 1200,
                      "subtotal_cents": 2400
                    },
                    {
                      "sku": "MUG-001",
                      "name": "Stoneware Mug",
                      "quantity": 1,
                      "unit_price_cents": 2400,
                      "subtotal_cents": 2400
                    }
                  ],
                  "shipping_address": {
                    "line1": "1 Analytical Engine Way",
                    "city": "London",
                    "country": "GB"
                  }
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "errors": {
                      "type": "object",
                      "properties": {
                        "line_items": {
                          "type": "array",
                          "items": {
                            "type": "string",
                            "example": "can't be blank"
                          }
                        },
                        "sku": {
                          "type": "array",
                          "items": {
                            "type": "string",
                            "example": "TEA-999 is not a known product"
                          }
                        }
                      }
                    }
                  }
                },
                "example": {
                  "errors": {
                    "line_items": [
                      "can't be blank"
                    ]
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v2/products": {
      "get": {
        "operationId": "get_api_v2_products",
        "summary": "List Products",
        "description": "api/v2/products#index\nRoute: api_v2_products",
        "tags": [
          "Products"
        ],
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "example": "eyJpZCI6Mn0"
            },
            "example": "eyJpZCI6Mn0"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer",
                            "example": 2
                          },
                          "sku": {
                            "type": "string",
                            "example": "TEA-002"
                          },
                          "name": {
                            "type": "string",
                            "example": "Sencha"
                          },
                          "category": {
                            "type": "string",
                            "enum": [
                              "mugs",
                              "tea"
                            ],
                            "example": "tea"
                          },
                          "price_cents": {
                            "type": "integer",
                            "example": 1500
                          },
                          "tags": {
                            "type": "array",
                            "items": {
                              "type": "string",
                              "enum": [
                                "green",
                                "insulated"
                              ],
                              "example": "green"
                            }
                          },
                          "price": {
                            "type": "string",
                            "example": "15.00"
                          }
                        },
                        "required": [
                          "id",
                          "sku",
                          "name",
                          "category",
                          "price_cents",
                          "tags",
                          "price"
                        ]
                      }
                    },
                    "next_cursor": {
                      "type": [
                        "null",
                        "string"
                      ],
                      "example": "eyJpZCI6Mn0"
                    }
                  },
                  "required": [
                    "items",
                    "next_cursor"
                  ]
                },
                "example": {
                  "items": [
                    {
                      "id": 1,
                      "sku": "TEA-001",
                      "name": "Earl Grey",
                      "category": "tea",
                      "price_cents": 1200,
                      "tags": [
                        "black",
                        "bergamot"
                      ],
                      "price": "12.00"
                    },
                    {
                      "id": 2,
                      "sku": "TEA-002",
                      "name": "Sencha",
                      "category": "tea",
                      "price_cents": 1500,
                      "tags": [
                        "green"
                      ],
                      "price": "15.00"
                    }
                  ],
                  "next_cursor": "eyJpZCI6Mn0"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key"
      }
    }
  }
}